back/api/requirements.txt
gunicorn>=20.1.0
Django>=4.0,<5
psycopg2-binary>=2.8,<4
djangorestframework>=3.13,<4
drf-extensions>=0.6,<1
drf-yasg>=1.20,<2
drf-nested-routers>=0.92,<1
mysqlclient>=2.0,<3
PyJWT>=2.3,<3
djangorestframework-simplejwt>=5.1,<6
django-cors-headers>=3.11,<4
django-rest-passwordreset>=1.2,<2
djangorestframework-mvt>=0.2,<1
python-dotenv>=0.20<1
pyodbc>=4.0.32
psycopg2-binary>=2.9.3
django-multiselectfield>=0.1.12
aiohttp>=3.8.1
requests-futures>= 1.0.0
whitenoise>=6.4.0
routingpy>=1.2.1
pillow>=10.0.0
redis>=5.0.1